FUSETOOLS2-2605: workaround for codelenses that could be recomputed
sometimes, the codelens elements are recomputed between the time we find
the element and the time we try to click on it. it leads to stale
reference error.
Applied a workaround which is consists in retrying one time in case it
happens. It was successful 2 times in row on CI for all jobs (last
builds without the workaround, it was always failing at least one of the
matrix job)

Signed-off-by: Aurélien Pupier <apupier@redhat.com>