Use Camel 4.9.0 by default for Camel JBang
By default, Camel Jbang is now using Jib and not Docker and providing a
multi-arch platform and the way to produce images are no more the same
consequently:
* a specific registry must be configured. We found only using Docker to
make it working with Minikube.
* qemu is required for the multi-arch with Docker
* resgitry-alias is required because vs code tasks cannot escape and
interpret correctly the env command of the registry setup

note that it is fixing the deploy only.

The debug (and so remaining test failures) is not working due to
https://issues.apache.org/jira/browse/CAMEL-21612

Signed-off-by: Aurélien Pupier <apupier@redhat.com>