Use Camel 4.10.2 by default for Camel JBang
By default, Camel Jbang is now using Jib and not Docker and providing a
multi-arch platform and the way to produce images are no more the same
consequently:
* a specific registry must be configured. We found only using Docker to
make it working with Minikube.
* qemu is required for the multi-arch with Docker
* registry-alias is required because vs code tasks cannot escape and
interpret correctly the env command of the registry setup
* the build is now launched by default with `--quiet` option, so there
is no more BUILD SUCCESS visible

note that it is fixing the deploy only.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>